Don't install msi for base R3.15
from:anj@aps.anl.gov

Base-3.15 compiles its own version of msi which has some additional
features. Please switch to using that instead of the copy from the
extensions tree, which cannot be used with the 3.15 build rules anyway
(which require support for a -D flag to have it output dependency data).
I have deliberately not published a version of msi that builds against
base-3.15.