Add option '-i' to `install-epics.sh` for automatic installation
install-epics.sh has now the option to run in a non-interactive mode,
which is useful for automated installations.

The script can now be called with an argument `-i y|Y`.
If this argument is present, its value is used as an answer to the
interactive question whether epics should be installed or not.

Examples:
 - `./install-epics.sh` - script asks user interactively
 - `./install-epics.sh -i y` - EPICS is installed automatically, no interaction required
 - `./install-epics.sh -i n` - EPICS is not installed, no interaction required