compile-epics.sh: EPICS_DEBUG did not work
Compiling EPICS with DEBUG (no optimize, debug symbols) did not work.
The file CONFIG_gnuCommon was patched with a patchfile dependend on
the version of EPICS base.
The `patch` program was unable to find the patch file, since it
is in the root directory of our project.
Solution: Copy it from the right place