Add compilation of seq; partly fix re2c
Make it possible to install re2c ?
Fix the original code a litte bit - the one that compiles re2c using git
clone and then ./autogen.sh && ./configure.sh && ...

However, manual intervention may be needed still.
Or there may be a package for the Linux (and other) systems.
However, since this needs to be done only if the sequencer is needed,
it is hard to re-test on the same machine.